break out classes into files in experiment_box
