+++ pytc_gui/widgets/experiment_box/experiment_dialog/__init__.py
@@ -1,0 +1,8 @@
+__description__ = \
+"""
+Experiment fit parameter and properties dialog box.
+"""
+__author__ = "Michael J. Harms"
+__date__ = "2017-06-01"
+
+from .experiment_dialog import ExperimentOptionsDialog
