+++ pytc_gui/widgets/experiment_box/experiment_dialog/connector_dialog.py
