+++ pytc_gui/widgets/experiment_box/experiment_dialog/fit_param_wrapper.py
@@ -1,5 +1,6 @@
 __description__ = \
 """
+Create GUI elements wrapping a single fit parameter.
 """
 __author__ = "Michael J. Harms"
 __date__ = "2017-06-01"
@@ -7,10 +8,7 @@ __date__ = "2017-06-01"
 from .connector_dialog import AddConnectorDialog
 
 from PyQt5 import QtWidgets as QW
-from PyQt5 import QtGui as QG
 from PyQt5 import QtCore as QC
-
-import sys, copy
 
 class FitParamWrapper(QW.QWidget):
     """
@@ -323,249 +321,5 @@ class FitParamWrapper(QW.QWidget):
     @property
     def alias_widget(self):
         return self._alias 
-
-
-class ExperimentSettableWrapper(QW.QWidget):
-    """
-    Wrap a single experiment settable as a widget.
-    """
-
-    def __init__(self,parent,fit,experiment,settable_name,start_value,value_type,
-                 allowable_values,float_view_cutoff=100000):
-        """
-        parent: parent widget
-        fit: FitContainer object
-        experiment: pytc.ITCExperiment object containing settable
-        settable_name: name of settable (string) in experiment object
-        start_value: starting value of settable when widget opens
-        value_type: value type. 
-            bool -> QCheckBox
-            str,float,int -> QLineEdit, parsed appropriately
-            multi -> QDropDown, using allowable_values to populate
-        allowable_value: list of available values for multi, None otherwise
-        float_view_cutoff: how to show floats in QLineEdit boxes
-        """
-
-        super().__init__()
-
-        self._parent = parent
-        self._fit = fit
-        self._experiment = experiment
-        self._settable_name = settable_name
-        self._start_value = start_value
-        self._value_type = value_type
-        self._allowable_values = allowable_values
-        self._float_view_cutoff = float_view_cutoff
-
-
-        self.layout()
-
-    def layout(self):
-        """
-        Create widget layout.
-        """
-
-        self._main_layout = QW.QHBoxLayout(self)
-
-        self._label = QW.QLabel(self._settable_name)
-        self._main_layout.addWidget(self._label)       
-
-        # --------- multi value ----------
-        if self._value_type == "multi":
-
-            self._select_widget = QW.QComboBox(self)
-
-            try:
-                len(self._allowable_values)
-            except TypeError:
-                err = "allowable types must be a list-like object for a multi value.\n"
-                raise ValueError(err) 
-
-            self._allowable_values = copy.copy(self._allowable_values)
-            self._allowable_values.sort()
-            self._select_widget.addItems(self._allowable_values)
-
-            current_index = self._select_widget.findText(self._start_value)
-            if current_index == -1:
-                err = "current value is not in the allowable values\n"
-                raise ValueError(err)
-
-            self._select_widget.setCurrentIndex(current_index)
-            self._select_widget.currentIndexChanged.connect(self._update_multi_settable)
-
-        # --------- bool value --------------
-        elif self._value_type == bool:
-
-            self._select_widget = QW.QCheckBox()
-            self._select_widget.setChecked(self._start_value)
-            self._select_widget.stateChanged.connect(self._start_value)
-
-        # -------- other values --------------
-        else:
-            self._select_widget = QW.QLineEdit(self)
-
-            if self._value_type == float:
-                if self._start_value < 1/self._float_view_cutoff or self._start_value > self._float_view_cutoff:
-                    val_str = "{:.8e}".format(self._start_value)
-                else:
-                    val_str = "{:.8f}".format(self._start_value)
-            else:
-                val_str = "{}".format(self._start_value)
-
-            self._select_widget.setText(val_str)
-
-            self._select_widget.textChanged.connect(self._update_other_settable)
-
-        self._main_layout.addWidget(self._select_widget)
-           
-    def _update_multi_settable(self,value):
-        """
-        Handler for "multi" values.  Parses QDropDown.
-        """
-
-        new_value = self._select_widget.currentText()
-
-        setattr(self._experiment, self._settable_name, new_value)
-        if self._fit.continuous_update:
-            self._fit.emit_changed()
-
-    def _update_bool_settable(self):
-        """
-        Handler for bool values.  Parses QCheckBox
-        """
-              
-        setattr(self._experiment,self._settable_name, self._select_wiget.checkState())
-        if self._fit.continuous_update:
-            self._fit.emit_changed() 
-
-    def _update_other_settable(self):
-        """
-        Handler for other values (str, float, int most likely).  Parses
-        QLineEdit.
-        """
-
-        current = self._select_widget.text()
-        try:
-            new_value = self._value_type(current)
-            color = "#FFFFFF"
-        except ValueError:
-            new_value = None
-            color = "#FFB6C1"
-        
-        self._select_widget.setStyleSheet("QLineEdit {{ background-color: {} }}".format(color))
-        if new_value is not None:
-        
-            setattr(self._experiment,self._settable_name, new_value)
-            if self._fit.continuous_update:
-                self._fit.emit_changed() 
- 
-class ExperimentOptionsDialog(QW.QDialog):
-    """
-    Dialog used to create options for experiment.  This includes fit parameters
-    and some experiment metadata.
-    """
-
-    def __init__(self,parent,fit,experiment):
-        """
-        parent: parent widget
-        fit: FitContainer instance
-        experiment: pytc.ITCExperimentInstance
-        """
-
-        super().__init__()
-
-        self._parent = parent
-        self._fit = fit
-        self._experiment = experiment
-
-        # Use the FitContainer instance to get the fittable parameteres and the
-        # expeirment properties that can be set.
-        self._fit_param_info = self._fit.get_experiment_param(self._experiment)
-        self._experiment_settable = self._fit.get_experiment_settable(self._experiment)
-
-        self.layout()
-
-    def layout(self):
-        """
-        Construct the widget.
-        """
-    
-        model_params = list(self._fit_param_info.keys())
-        model_name = model_params.remove("model_name")
-        model_params.sort()
-
-        self._main_layout = QW.QVBoxLayout(self)
-       
-        # ------------ Fit parameters --------------- 
-        self._fit_param_box = QW.QGroupBox("Fit Parameters") 
-        self._fit_param_layout = QW.QGridLayout()
-
-        # Table header 
-        self._fit_param_layout.addWidget(QW.QLabel("param"), 0,0) 
-        self._fit_param_layout.addWidget(QW.QLabel("guess"), 0,1) 
-        self._fit_param_layout.addWidget(QW.QLabel("global"),0,2) 
-        self._fit_param_layout.addWidget(QW.QLabel("fixed"), 0,3) 
-        self._fit_param_layout.addWidget(QW.QLabel("lower"), 0,4) 
-        self._fit_param_layout.addWidget(QW.QLabel("upper"), 0,5) 
-
-        # Build widgets for each fittable parameter
-        self._param_widgets = []           
-        for i, k in enumerate(model_params):
-
-            self._param_widgets.append(FitParamWrapper(self,self._fit,
-                                                       self._experiment,
-                                                       self._fit_param_info[k]))
- 
-            self._fit_param_layout.addWidget(QW.QLabel(k),                        i+1,0)
-            self._fit_param_layout.addWidget(self._param_widgets[-1].guess_widget,i+1,1)
-            self._fit_param_layout.addWidget(self._param_widgets[-1].alias_widget,i+1,2)
-            self._fit_param_layout.addWidget(self._param_widgets[-1].fixed_widget,i+1,3)
-            self._fit_param_layout.addWidget(self._param_widgets[-1].lower_widget,i+1,4)
-            self._fit_param_layout.addWidget(self._param_widgets[-1].upper_widget,i+1,5)
-
-        self._fit_param_box.setLayout(self._fit_param_layout)
-        self._main_layout.addWidget(self._fit_param_box)
-
-        # ---------- Horizontal line -----------------
-        hline = QW.QFrame()
-        hline.setFrameShape(QW.QFrame.HLine)
-        self._main_layout.addWidget(hline)
-
-        # ---------- Experiment information ------------      
-
-        expt_keys = list(self._experiment_settable.keys())
-        expt_keys.sort()
- 
-        self._expt_widgets = []       
-        for settable_name in expt_keys:
-
-            start_value      = self._experiment_settable[settable_name][0]
-            value_type       = self._experiment_settable[settable_name][1]
-            allowable_values = self._experiment_settable[settable_name][2] 
-
-            self._expt_widgets.append(ExperimentSettableWrapper(self,
-                                                                self._fit,
-                                                                self._experiment,
-                                                                settable_name,
-                                                                start_value,
-                                                                value_type,
-                                                                allowable_values))
- 
-        self._experiment_settable_box = QW.QGroupBox("Experiment information") 
-        self._experiment_settable_layout = QW.QGridLayout()
-
-        # Layout the experimental widgets in rows of 3.
-        while len(self._expt_widgets) % 3 != 0:
-            self._expt_widgets.append(QW.QLabel(" "))
-
-        counter = 0
-        num_rows = int(round((len(self._expt_widgets)+1)/3))
-        for i in range(num_rows):
-            for j in range(3):
-                self._experiment_settable_layout.addWidget(self._expt_widgets[counter],i,j)
-                counter += 1
-
-        self._experiment_settable_box.setLayout(self._experiment_settable_layout)
-        self._main_layout.addWidget(self._experiment_settable_box)
